requirements.txt
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
alpaca-trade-api>=3.0.0
python-dotenv>=1.0.0 